Substitute synth_piece() > rt_synthesis() > dn_synthesis() by a much shorter and faster subroutine direct_synthesis() > synth_methods(), now without using pointers nor structures, and passing all optical profiles with dimensions in the order (w,z,stoks). This should maximize eficiency for integrating along space the RTE later. For this, all the optical coeffcients in the python code were also re-estructured. The first set of above subroutines were left for testing and comparison. They also required some adaptation to enter the profiles with the new dimensions but internally they work with the old ordering of dimensions.
c_rtcoeffs() now creates outputs directly with the new ordering in the dimensions of the optical coefficients.

Simplify implementation of the EvolOp RT method  to just a few lines in example subroutine synth_evolop(). synth_methods() is still main RT subroutine but we keep track of these auxiliary older subroutines.

Remove spectral convolution from synth_methods() (it is considered in Python). Such convolution was being done for every layer in Hazel, which was incorrect when there is more than one atmosphere cell in the RT. It  only makes sense at the end, for the outgoing stokes profile.

Abstract factor beta out of the RT method because it does not really belong to it. Now it multiplies the emissivity before calling the new RT routines in Python. It could also be moved to the end of the calculation of the emissivity in Fortran.

Add routine get_dsLOS() to calculate all height scale intervals (ds) of the model from Python including LOS effect (mu). This calculation can now be done from the dtau associated to each point, or from the user defined hz.

In ModelRT only, the keyword atom was deleted from add_spectrum and is now calculated more efficiently from lineHazel internally using a dictionary transforming from line to atom dic[line]=atom. Keyword lineHazel was changed to line. This could/should be repeated for Model.py.

Choice for plotting the optical coefficients centered around line center.

Separate the deallocation part in fortran into two subroutines that can now be called independently for deallocating outer or inner variables (i.e. those initialized in the main fortran program (hazel_py.f90) or in the inner subroutines). The subroutines involved in allocation and deallocation are alloc_outer_coefs() in rt_coefs.f90 and deallocate_outervars(), deallocate_innervars(), and deallocate_all() in hazel_py.f90.

Extract identity matrix to hazel_py.f90 to avoid repeating it for every cell.

Removal some unnecessary long names in structures of original routines.

BUG: correct bug in definition of indices segmenting the ray in pieces.
BUG:correct segmentation fault due to incorrect specifications in pyx file.
TEST: results with different subroutines and methods.
TEST: memory management (fine) and speed of calculations (correct). Typical times for solution of SEE and coeffs is 0.7 seconds for Sodium atom, 1 ms/ray for old RT subroutine, and 0.4 ms/ray for the new subroutines without yet adding Magnus scheme.
TEST: window and plotting management
CLEANING OF COMMENTS PENDING