Implement linear tau and plot grids of polynomials in tau.
Modification of synthesize():
 1)return an independent copy of the computed Stokes profiles for analysis/comparison.
2)compute directly the desired spectrum (a trivial synthesize_all() could be implemented later if necessary) instead of checking them all.
3) Move checking of filling factors to setup() for efficiency (anyway in this version we deactivated filling factors for consistency with the new methods).
4)Improve treatment of reading of model with FtR option and rescaling of the ghost figure f4 when loading a saved model. Add set_fontsize() and rescale figure() functions to re scale the figure.
5)Verify smooth saving and loading of model after solving the SEEs.

Remove PSF and init_psf() from fortran synth.f90 because it is not correct to convolve with a spectral PSF during the RT, only at the end of it. Now it can be done only from python.

Invert stokesIn and StokesOut dimensions to be Fortran efficient in direct_synthesis() subroutine as done in synth_piece() > rt_synthesis() > dn_synthesis(). Rest of synthesis routines preserve the old form to assure some backward compatibility in case is necessary and for being used by Model.py.

Verify calculations are right when calling synthesize repeatedly.

Program plot_coeffs2D()

Add legend in fig1 for the different methods
Add line center and bandwidth lines in figures 1 and 5