Implement and verify our Magnus-1 method with our second alternative expression.
Implement Magnus-2, which includes second order accuracy correction to Magnus expansion with our expression with 5-order precision.

BUG:Correct Magnus expressions (typo in Lorentz functions)

Implementation of DELO, DELOPAR, DELOPARABOLLIC,  and trapezoidal standard method without simplifying exponentials with power series. This latter practice would kill the gruop structure and by avoiding it we are already improving accuracy of these standard methods.

BUG: Correct -2 signs in DELO-parabollic expression.

Preliminar tests of methods apparently working fine.

Increment number of possible atmospheric points in Fortran structures used to calculate the SEE and the optical coefficients. These structures are obsolete in the new ModelRT.py model and in the new synthesize routines but we keep them for backward compatibility. They were limited to 100 points and we set them now to 500, but the number of points could be dynamically read wheh initializing Hazel. If this number is not adapted, segmentation faults will occur.

Label methods correctly in Python dictionaries and in Fortran source code.